Moves pseudo-block GMRES data structures to iteration header
Also transitions pseudo block GMRES to use dense matrix traits
